rlib: move rbigint's storage and GC vocabulary out of the algorithm port
`rpython/rlib/rbigint.py` names no malloc, array type id, root or write
barrier. The storage words belong to the rtyper (`ll_newlist`,
rtyper/lltypesystem/rlist.py:324-329; `ll_alloc_and_set`, rtyper/rlist.py:494-503)
and the GC words are generated by the transform
(`rpython/memory/gctransform/framework.py`), which runs two pipeline stages
after the codewriter builds jitcodes (`rpython/translator/driver.py:348` vs
`:408`). pyre's port carried 42 such mentions in the algorithm file.

Two relocations, no behaviour change:

`Digits` in `lltypesystem/rlist.rs` — `new`/`try_new` (`ll_newlist`),
`alloc_and_set_zero`/`try_alloc_and_set_zero`
(`ll_alloc_and_set(LIST, count, 0)`, i.e. `[NULLDIGIT] * count`), and
`prebuilt`. The six construction sites in `rbigint.rs` called
`alloc_typed_items_block_nursery`, `try_alloc_typed_items_block_nursery`,
`alloc_typed_items_block_immortal` and `typed_items_block_clear` with
`GC_INT_ARRAY_GC_TYPE_ID`; they now call `Digits`, and the array's type id and
allocation tier are spelled only in the rlist layer.

`rbigint/gc.rs` — 430 lines moved: `RBigIntGcRoot`, `RBIGINT_DIGITS_OFFSET`,
`RBIGINT_PAYLOAD_SIZE`, the payload and `tuple2` GC type ids and their setters,
`prebuilt_payload_pointer`, the seven `alloc_rbigint_*` payload allocators,
`PendingPairItemRoot` and the two pair allocators,
`PendingPartsCacheDigitRoot`, and `walk_rbigint_cache_digit_slots`.
`pub use gc::*` keeps all 29 external `majit_rlib::rbigint::*` spellings.

Production code in `rbigint.rs` now counts zero for `majit_gc`,
`GcAllocOutcome`, `gc_add_root`, `gc_remove_root`, `gc_write_barrier`,
`alloc_rbigint_*`, `majit_ir::GcRef`, `*_GC_TYPE_ID`, `nursery`, `oldgen`,
`immortal` and `malloc_raw`; four mentions remain in `#[cfg(test)]` allocator
tests.

Three things stayed, each because something keys on where they are:

- The `RBigIntPair` struct and its offset constants.
  `pyre-jit-trace/src/descr.rs:1823` keys its descr group on the def-path
  string `"rbigint::RBigIntPair"`, with no test pinning it. Only its
  allocators moved.
- `RBigInt._digits`'s type. `front/mir.rs:1405` overrides that field to
  `[i64]` for the annotator and `test_rbigint_mir.rs:704` pins `RBigInt::digit`
  to a direct `_digits` projection, so a newtype's `.0` would add an op between
  them. `Digits` is a namespace, not a wrapper.
- The read path. `graph_is_items_block_base_accessor` (`front/mir.rs:17258`)
  matches `rlist::typed_items_block_items_base` by path, and the
  `.add(ITEMS_OFFSET)` fold depends on it.
